Moving Access enum class into types subfolder.
