+++ src/register/RegisterDerived.hpp
@@ -13,8 +13,7 @@
  **************************************************************************************/
 
 #include "RegisterBase.h"
-#include "RegisterAccess.h"
-
+#include "types/Access.h"
 #include "util/register_util.hpp"
 
 /**************************************************************************************
